requirements.txt
chromadb
openai
python-dotenv
sentence-transformers